requirements.txt
asyncpg==0.22.0
fastapi>=0.68.0,<0.69.0
psycopg2-binary==2.8.6
SQLAlchemy==1.4.0
uvicorn>=0.15.0,<0.16.0
requests>=2.26.0,<2.27.0
python-dotenv==0.20.0
